GV-339 moved the loader of chart in new component
