+++ packages/manager/src/pages/dashboard/components/dashboard-portfolio-chart-section/dashboard-chart/dashboard-portfolio-chart-container.js
@@ -2,8 +2,10 @@ import React, { Fragment, PureComponent } from "react";
 import { connect } from "react-redux";
 import { bindActionCreators } from "redux";
 import ChartPeriod from "shared/components/chart/chart-period/chart-period";
-import DashboardChartLoader from "shared/components/dashboard/dashboard-chart-loader/dashboard-chart-loader";
-import { DashboardChartDescriptionLoader } from "shared/components/dashboard/dashboard-chart-loader/dashboard-chart-loaders";
+import {
+  DashboardChartDescriptionLoader,
+  DashboardChartLoader
+} from "shared/components/dashboard/dashboard-chart-loader/dashboard-chart-loaders";
 import FundProfitChart from "shared/components/funds/fund-details/fund-details-statistics-section/fund-details-chart-section/fund-profit-chart-section/fund-profit-chart";
 import ProgramProfitChart from "shared/components/programs/program-details/program-details-statistic-section/program-details-chart-section/program-profit-chart-section/program-profit-chart";
